requirements.txt
debugpy==1.6.5
comm==0.1.2
ipython==8.10.0
pycaret
mrmr_selection
pandas 
seaborn
numpy
lightgbm
deepchecks==0.12.0
xgboost
catboost
# skleran